requirements.txt
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
scipy==1.11.4
tqdm==4.66.4
requests==2.31.0
joblib==1.4.0
typer==0.9.0
streamlit==1.31.0
matplotlib==3.8.3
seaborn==0.13.2
pyyaml==6.0.1
pytest==8.1.1

